llm_learning/intro-llm-code/chs/ch4-distributed/DistributedDataParallel/requirements.txt
python==3.10.16
torch==2.6.0+cu118
torchvision==0.21.0+cu118
numpy==2.1.2
pillow==11.0.0